Move AssemblyDropdown to ToolbarAssemblySection.cs
