+++ osu.Framework/Testing/Drawables/Sections/ToolbarAssemblySection.cs
@@ -1,7 +1,6 @@
 // Copyright (c) 2007-2018 ppy Pty Ltd <contact@ppy.sh>.
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu-framework/master/LICENCE
 
-using System.Linq;
 using System.Reflection;
 using osu.Framework.Allocation;
 using osu.Framework.Graphics;
@@ -61,5 +60,20 @@ namespace osu.Framework.Testing.Drawables.Sections
         }
 
         public void AddAssembly(string name, Assembly assembly) => assemblyDropdown.AddAssembly(name, assembly);
+
+        private class AssemblyDropdown : BasicDropdown<Assembly>
+        {
+            public void AddAssembly(string name, Assembly assembly)
+            {
+                if (assembly == null) return;
+                foreach(var item in MenuItems)
+                {
+                    if(item.Text.Value.Contains("dynamic"))
+                        RemoveDropdownItem(item.Value);
+                }
+
+                AddDropdownItem(name, assembly);
+            }
+        }
     }
 }
